Make scaleDownCluster stage support providers selection and implement for Google. Existing implementation was massaged into AWS-specific implementation.
