+++ app/scripts/modules/core/pipeline/config/stages/scaleDownCluster/aws/awsScaleDownClusterStage.js
@@ -1,0 +1,83 @@
+'use strict';
+
+let angular = require('angular');
+
+module.exports = angular.module('spinnaker.core.pipeline.stage.aws.scaleDownClusterStage', [
+  require('../../../../../utils/lodash.js'),
+  require('../../stageConstants.js'),
+  require('./scaleDownClusterExecutionDetails.controller.js')
+])
+  .config(function(pipelineConfigProvider) {
+    pipelineConfigProvider.registerStage({
+      provides: 'scaleDownCluster',
+      cloudProvider: 'aws',
+      templateUrl: require('./scaleDownClusterStage.html'),
+      executionDetailsUrl: require('./scaleDownClusterExecutionDetails.html'),
+      validators: [
+        { type: 'requiredField', fieldName: 'cluster' },
+        { type: 'requiredField', fieldName: 'remainingFullSizeServerGroups', fieldLabel: 'Keep [X] full size Server Groups'},
+        { type: 'requiredField', fieldName: 'regions', },
+        { type: 'requiredField', fieldName: 'credentials', fieldLabel: 'account'},
+      ],
+      strategy: true,
+    });
+  }).controller('awsScaleDownClusterStageCtrl', function($scope, accountService, stageConstants, _) {
+    var ctrl = this;
+
+    let stage = $scope.stage;
+
+    $scope.state = {
+      accounts: false,
+      regionsLoaded: false
+    };
+
+    accountService.listAccounts('aws').then(function (accounts) {
+      $scope.accounts = accounts;
+      $scope.state.accounts = true;
+    });
+
+    $scope.regions = ['us-east-1', 'us-west-1', 'eu-west-1', 'us-west-2'];
+
+    ctrl.accountUpdated = function() {
+      accountService.getRegionsForAccount(stage.credentials).then(function(regions) {
+        $scope.regions = _.map(regions, function(v) { return v.name; });
+        $scope.state.regionsLoaded = true;
+      });
+    };
+
+    stage.regions = stage.regions || [];
+    stage.cloudProvider = 'aws';
+
+    if (!stage.credentials && $scope.application.defaultCredentials) {
+      stage.credentials = $scope.application.defaultCredentials;
+    }
+    if (!stage.regions.length && $scope.application.defaultRegion) {
+      stage.regions.push($scope.application.defaultRegion);
+    }
+
+    if (stage.credentials) {
+      ctrl.accountUpdated();
+    }
+
+    if (stage.remainingFullSizeServerGroups === undefined) {
+      stage.remainingFullSizeServerGroups = 1;
+    }
+
+    if (stage.allowScaleDownActive === undefined) {
+      stage.allowScaleDownActive = false;
+    }
+
+    ctrl.pluralize = function(str, val) {
+      if (val === 1) {
+        return str;
+      }
+      return str + 's';
+    };
+
+    if (stage.preferLargerOverNewer === undefined) {
+      stage.preferLargerOverNewer = "false";
+    }
+    stage.preferLargerOverNewer = stage.preferLargerOverNewer.toString();
+  })
+  .name;
+
