+++ app/scripts/modules/core/pipeline/config/stages/scaleDownCluster/aws/scaleDownClusterStage.html
@@ -1,0 +1,56 @@
+<div ng-controller="awsScaleDownClusterStageCtrl as scaleDownClusterStageCtrl" class="form-horizontal">
+  <div ng-if="!pipeline.strategy">
+    <div class="form-group">
+      <label class="col-md-2 col-md-offset-1 sm-label-left">Account</label>
+      <div class="col-md-6">
+        <account-select-field component="stage" field="credentials" accounts="accounts" provider="'aws'" on-change="scaleDownClusterStageCtrl.accountUpdated()" required></account-select-field>
+      </div>
+    </div>
+    <div class="form-group">
+      <label class="col-md-2 col-md-offset-1 sm-label-left">Regions</label>
+      <div class="col-md-6" ng-if="!stage.credentials">
+        <p class="form-control-static">(Select an Account)</p>
+      </div>
+      <div class="col-md-6" ng-if="stage.credentials">
+        <checklist items="regions" model="stage.regions" inline="true" include-select-all-button="true"></checklist>
+      </div>
+    </div>
+    <div class="form-group">
+      <div class="col-md-2 col-md-offset-1 sm-label-left">
+        <b>Cluster</b>
+        <help-field key="pipeline.config.scaleDownCluster.cluster"></help-field>
+      </div>
+      <div class="col-md-6">
+        <input type="text" required ng-model="stage.cluster"
+          class="form-control input-sm" />
+      </div>
+    </div>
+  </div>
+  <div class="form-group">
+    <div class="col-md-2 col-md-offset-1 sm-label-left">
+      <b>Scale Down Options</b>
+    </div>
+    <div class="col-md-6">
+      <div class="form-inline">
+        <p>Keep the
+        <input type="number" min="0" required ng-model="stage.remainingFullSizeServerGroups"
+               class="form-control input-sm" style="width: 50px" />
+        <select class="form-control input-sm"
+                ng-model="stage.preferLargerOverNewer" style="width: 100px">
+          <option value="true">largest</option>
+          <option value="false">newest</option>
+        </select>
+        {{scaleDownClusterStageCtrl.pluralize('server group', stage.remainingFullSizeServerGroups)}} at current size.</p>
+        <p>The remaining server groups will be scaled down to zero instances.</p>
+      </div>
+    </div>
+  </div>
+  <div class="form-group">
+    <div class="col-md-offset-3 col-md-6 checkbox">
+      <label>
+        <input type="checkbox" ng-model="stage.allowScaleDownActive"/>
+        Allow scale down of active server groups
+      </label>
+    </div>
+  </div>
+</div>
