+++ app/scripts/modules/core/pipeline/config/stages/scaleDownCluster/gce/scaleDownClusterExecutionDetails.html
@@ -1,4 +1,4 @@
-<div ng-controller="ScaleDownClusterExecutionDetailsCtrl">
+<div ng-controller="gceScaleDownClusterExecutionDetailsCtrl">
     <execution-details-section-nav sections="configSections"></execution-details-section-nav>
     <div class="step-section-details" ng-if="detailsSection === 'scaleDownClusterConfig'">
         <div class="row">
@@ -8,8 +8,8 @@
                     <dd>
                         <account-tag account="stage.context.credentials" provider="stage.context.cloudProvider"></account-tag>
                     </dd>
-                    <dt>Region</dt>
-                    <dd>{{stage.context.regions.join(', ')}}</dd>
+                    <dt>Zone</dt>
+                    <dd>{{stage.context.zones.join(', ')}}</dd>
                     <dt>Cluster</dt>
                     <dd>{{stage.context.cluster}}</dd>
                     <dt>Keep Full Size</dt>
@@ -24,7 +24,7 @@
             <div class="col-md-11">
                 <dl class="dl-narrow dl-horizontal">
                     <dt>Scaled down:</dt><dd></dd>
-                    <dt ng-repeat-start="(region, servergroups) in stage.context['deploy.server.groups']">{{region}}</dt>
+                    <dt ng-repeat-start="(zone, servergroups) in stage.context['deploy.server.groups']">{{zone}}</dt>
                     <dd ng-repeat-end>{{servergroups.join(', ')}}</dd>
                 </dl>
             </div>
