+++ app/scripts/modules/core/pipeline/config/stages/scaleDownCluster/gce/scaleDownClusterStage.html
@@ -1,18 +1,18 @@
-<div class="form-horizontal">
+<div ng-controller="gceScaleDownClusterStageCtrl as scaleDownClusterStageCtrl" class="form-horizontal">
   <div ng-if="!pipeline.strategy">
   <div class="form-group">
     <label class="col-md-2 col-md-offset-1 sm-label-left">Account</label>
     <div class="col-md-6">
-      <account-select-field component="stage" field="credentials" accounts="accounts" provider="'aws'" on-change="scaleDownClusterStageCtrl.accountUpdated()" required></account-select-field>
+      <account-select-field component="stage" field="credentials" accounts="accounts" provider="'gce'" on-change="scaleDownClusterStageCtrl.accountUpdated()" required></account-select-field>
     </div>
   </div>
   <div class="form-group">
-    <label class="col-md-2 col-md-offset-1 sm-label-left">Regions</label>
+    <label class="col-md-2 col-md-offset-1 sm-label-left">Zones</label>
     <div class="col-md-6" ng-if="!stage.credentials">
       <p class="form-control-static">(Select an Account)</p>
     </div>
     <div class="col-md-6" ng-if="stage.credentials">
-      <checklist items="regions" model="stage.regions" inline="true" include-select-all-button="true"></checklist>
+      <checklist items="zones" model="stage.zones" inline="true" include-select-all-button="false"></checklist>
     </div>
   </div>
   <div class="form-group">
